Establishing the game - Snake
In this part, the Assets folder was separated into Images and Prefabs;
images folder containing the borders of the game and also the food
image (normal small box). In the prefabs folder it contained the food
prefabs. I’ve created this prefab since the game will reproduce more
and more food constantly.