Adding Prefabs and scripts for food to spawn
FoodPrefab was added in order to recall the same script when adding the
foodPrefab to any sort of other script. Code was written and also
commented in detail. The code shows that the script will now spawn food
every 3 seconds starting from 2 seconds within the borders of the game
(Top, Bottom, Left & Right). Also the value was declared as and INT
value so that food is spawned within position (1,1) and not
(1.423,4.957).